Put utils to dedicated mod
